moved summary above events
