+++ public/components/common/search/search.tsx
@@ -7,9 +7,11 @@ import './search.scss';
 
 import '@algolia/autocomplete-theme-classic';
 import {
+  EuiAccordion,
   EuiBadge,
   EuiButton,
   EuiButtonEmpty,
+  EuiCallOut,
   EuiComboBox,
   EuiComboBoxOptionOption,
   EuiContextMenuItem,
@@ -17,8 +19,12 @@ import {
   EuiFlexGroup,
   EuiFlexItem,
   EuiIcon,
+  EuiLink,
+  EuiMarkdownFormat,
+  EuiPanel,
   EuiPopover,
   EuiPopoverFooter,
+  EuiSpacer,
   EuiSuperSelect,
   EuiSuperSelectOption,
   EuiText,
@@ -26,7 +32,7 @@ import {
 } from '@elastic/eui';
 import { isEqual } from 'lodash';
 import React, { useEffect, useState } from 'react';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { APP_ANALYTICS_TAB_ID_REGEX } from '../../../../common/constants/explorer';
 import { PPL_SPAN_REGEX } from '../../../../common/constants/shared';
 import { uiSettingsService } from '../../../../common/utils';
@@ -36,6 +42,7 @@ import { coreRefs } from '../../../framework/core_refs';
 import { SQLService } from '../../../services/requests/sql';
 import { SavePanel } from '../../event_analytics/explorer/save_panel';
 import { update as updateSearchMetaData } from '../../event_analytics/redux/slices/search_meta_data_slice';
+import { selectQueryAssistantSummarization } from '../../event_analytics/redux/slices/query_assistant_summarization_slice';
 import { PPLReferenceFlyout } from '../helpers';
 import { LiveTailButton, StopLiveButton } from '../live_tail/live_tail_button';
 import { Autocomplete } from './autocomplete';
@@ -44,6 +51,7 @@ import { QUERY_LANGUAGE } from '../../../../common/constants/data_sources';
 import { QueryArea } from './query_area';
 import './search.scss';
 import { useCatIndices, useGetIndexPatterns } from '../../event_analytics/explorer/llm/input';
+import chatLogo from '../../datasources/icons/query-assistant-logo.svg';
 export interface IQueryBarProps {
   query: string;
   tempQuery: string;
@@ -104,6 +112,7 @@ export const Search = (props: any) => {
     setIsQueryRunning,
   } = props;
 
+  const queryAssistantSummarization = useSelector(selectQueryAssistantSummarization)[tabId];
   const dispatch = useDispatch();
   const appLogEvents = tabId.match(APP_ANALYTICS_TAB_ID_REGEX);
   const [isSavePanelOpen, setIsSavePanelOpen] = useState(false);
@@ -448,18 +457,70 @@ export const Search = (props: any) => {
           </EuiFlexGroup>
         </EuiFlexItem>
         {!appLogEvents && (
-          <EuiFlexItem>
-            <QueryArea
-              tabId={tabId}
-              handleQueryChange={handleQueryChange}
-              handleTimeRangePickerRefresh={handleTimeRangePickerRefresh}
-              runQuery={query}
-              tempQuery={tempQuery}
-              setNeedsUpdate={setNeedsUpdate}
-              setFillRun={setFillRun}
-              selectedIndex={selectedIndex}
-            />
-          </EuiFlexItem>
+          <>
+            <EuiFlexItem>
+              <QueryArea
+                tabId={tabId}
+                handleQueryChange={handleQueryChange}
+                handleTimeRangePickerRefresh={handleTimeRangePickerRefresh}
+                runQuery={query}
+                tempQuery={tempQuery}
+                setNeedsUpdate={setNeedsUpdate}
+                setFillRun={setFillRun}
+                selectedIndex={selectedIndex}
+              />
+            </EuiFlexItem>
+            <EuiFlexItem grow={false}>
+              <EuiPanel>
+                <EuiAccordion
+                  id="summarization-accordion"
+                  buttonContent="AI Insights"
+                  initialIsOpen
+                  isLoading={queryAssistantSummarization?.summaryLoading ?? false}
+                  isLoadingMessage="Loading summary.."
+                  extraAction={
+                    <EuiFlexGroup direction="row" alignItems="center" gutterSize="s">
+                      <EuiFlexItem grow={false}>
+                        <EuiText color="subdued">
+                          <small>Generated by Opensearch Assistant</small>
+                        </EuiText>
+                      </EuiFlexItem>
+                      <EuiFlexItem grow={false}>
+                        <EuiIcon type={chatLogo} size="l" />
+                      </EuiFlexItem>
+                    </EuiFlexGroup>
+                  }
+                >
+                  {queryAssistantSummarization?.summary?.length > 0 && (
+                    <>
+                      <EuiSpacer size="m" />
+                      {queryAssistantSummarization?.isPPLError ? (
+                        <EuiCallOut title="There was an error" color="danger" iconType="alert">
+                          <EuiMarkdownFormat>
+                            {queryAssistantSummarization?.summary}
+                          </EuiMarkdownFormat>
+                        </EuiCallOut>
+                      ) : (
+                        <EuiPanel color="subdued" style={{ marginLeft: 16, marginRight: 16 }}>
+                          <EuiMarkdownFormat>
+                            {queryAssistantSummarization?.summary}
+                          </EuiMarkdownFormat>
+                        </EuiPanel>
+                      )}
+                      <EuiSpacer size="m" />
+                      <EuiText color="subdued">
+                        <small>
+                          The OpenSearch Assistant may produce inaccurate information. Verify all
+                          information before using it in any environment or workload. Share feedback
+                          via <EuiLink>Email</EuiLink> or <EuiLink>Slack</EuiLink>
+                        </small>
+                      </EuiText>
+                    </>
+                  )}
+                </EuiAccordion>
+              </EuiPanel>
+            </EuiFlexItem>
+          </>
         )}
       </EuiFlexGroup>
       {flyout}
