+++ public/components/event_analytics/explorer/explorer.tsx
@@ -103,7 +103,6 @@ import { selectQueryResult } from '../redux/slices/query_result_slice';
 import { changeDateRange, changeQuery, selectQueries } from '../redux/slices/query_slice';
 import { updateTabName } from '../redux/slices/query_tab_slice';
 import { selectExplorerVisualization } from '../redux/slices/visualization_slice';
-import { selectQueryAssistantSummarization } from '../redux/slices/query_assistant_summarization_slice';
 import {
   change as changeVisualizationConfig,
   change as changeVizConfig,
@@ -129,7 +128,6 @@ import {
   DEFAULT_DATA_SOURCE_TYPE,
   QUERY_LANGUAGE,
 } from '../../../../common/constants/data_sources';
-import chatLogo from '../../datasources/icons/query-assistant-logo.svg';
 
 export const Explorer = ({
   pplService,
@@ -184,7 +182,6 @@ export const Explorer = ({
   const explorerVisualizations = useSelector(selectExplorerVisualization)[tabId];
   const userVizConfigs = useSelector(selectVisualizationConfig)[tabId] || {};
   const explorerSearchMeta = useSelector(selectSearchMetaData)[tabId] || {};
-  const queryAssistantSummarization = useSelector(selectQueryAssistantSummarization)[tabId];
   const [selectedContentTabId, setSelectedContentTab] = useState(TAB_EVENT_ID);
   const [selectedCustomPanelOptions, setSelectedCustomPanelOptions] = useState([]);
   const [selectedPanelName, setSelectedPanelName] = useState('');
@@ -490,63 +487,6 @@ export const Explorer = ({
               <>
                 <EuiFlexItem grow={false}>
                   <EuiPanel hasBorder={false} hasShadow={false} paddingSize="s" color="transparent">
-                    <EuiPanel>
-                      <EuiAccordion
-                        id="summarization-accordion"
-                        buttonContent="Summary"
-                        initialIsOpen
-                        isLoading={queryAssistantSummarization?.summaryLoading ?? false}
-                        isLoadingMessage="Loading summary.."
-                        extraAction={
-                          <EuiFlexGroup direction="row" alignItems="center" gutterSize="s">
-                            <EuiFlexItem grow={false}>
-                              <EuiText color="subdued">
-                                <small>Generated by Opensearch Assistant</small>
-                              </EuiText>
-                            </EuiFlexItem>
-                            <EuiFlexItem grow={false}>
-                              <EuiIcon type={chatLogo} size="l" />
-                            </EuiFlexItem>
-                          </EuiFlexGroup>
-                        }
-                      >
-                        {queryAssistantSummarization?.summary?.length > 0 && (
-                          <>
-                            <EuiSpacer size="m" />
-                            {queryAssistantSummarization?.isPPLError ? (
-                              <EuiCallOut
-                                title="There was an error"
-                                color="danger"
-                                iconType="alert"
-                              >
-                                <EuiMarkdownFormat>
-                                  {queryAssistantSummarization?.summary}
-                                </EuiMarkdownFormat>
-                              </EuiCallOut>
-                            ) : (
-                              <EuiPanel color="subdued" style={{ marginLeft: 16, marginRight: 16 }}>
-                                <EuiMarkdownFormat>
-                                  {queryAssistantSummarization?.summary}
-                                </EuiMarkdownFormat>
-                              </EuiPanel>
-                            )}
-                            <EuiSpacer size="m" />
-                            <EuiText color="subdued">
-                              <small>
-                                The OpenSearch Assistant may produce inaccurate information. Verify
-                                all information before using it in any environment or workload.
-                                Share feedback via <EuiLink>Email</EuiLink> or{' '}
-                                <EuiLink>Slack</EuiLink>
-                              </small>
-                            </EuiText>
-                          </>
-                        )}
-                      </EuiAccordion>
-                    </EuiPanel>
-                  </EuiPanel>
-                </EuiFlexItem>
-                <EuiFlexItem grow={false}>
-                  <EuiPanel hasBorder={false} hasShadow={false} paddingSize="s" color="transparent">
                     {countDistribution?.data && !isLiveTailOnRef.current && (
                       <EuiPanel>
                         <HitsCounter
@@ -674,7 +614,6 @@ export const Explorer = ({
     query,
     isLiveTailOnRef.current,
     isQueryRunning,
-    queryAssistantSummarization,
   ]);
 
   const visualizations: IVisualizationContainerProps = useMemo(() => {
